move GTM file into its own folder
