+++ core/app/[locale]/(default)/_components/GTM/GTM.tsx
